Speed optimizations for new display driver.
This required to introduce a change in the internal API of the pixel_iterator.
Before it was possible to ask for a window area and not fill all of its
pixels. Now it is required to call the invalidate() member function of the
iterator in such cases. Note that since the pixel_iterator is not part of the
public API of the library, this is not a breaking change for applications.