Refactoring of font rendering engines + bugfix
* The 6 rendering engines have been merged into just 2 methods: one for
standard rendering and one for clipped rendering. These methods make
heavy use of templating in order to delegate  all the
burden of generating a different method depending on the font
type (different lookup and rendering logic of glyphs) to the
compiler.
It may be a little more cumbersome to read the method calls due to the
increased number of template parameters, but code
duplication is greatly reduced and mantainability is enhanced.

* Bugfix for FixedWidthCodeGenerator, it was printing the final codepoint of
the range instead of the range size. The MiscFixed header file has
been updated accordingly.

Signed-off-by: Federico Terraneo <fede.tft@miosix.org>